# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "BibleVerse": {
        "additionalProperties": true,
        "properties": {
          "book_english": {
            "title": "Book English",
            "type": "string"
          },
          "book_id": {
            "title": "Book Id",
            "type": "integer"
          },
          "book_zomi": {
            "title": "Book Zomi",
            "type": "string"
          },
          "changes_applied": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changes Applied"
          },
          "chapter": {
            "title": "Chapter",
            "type": "integer"
          },
          "english": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "English"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "testament": {
            "title": "Testament",
            "type": "string"
          },
          "verse": {
            "title": "Verse",
            "type": "integer"
          },
          "zomi_iso": {
            "title": "Zomi Iso",
            "type": "string"
          },
          "zomi_original": {
            "title": "Zomi Original",
            "type": "string"
          }
        },
        "required": [
          "testament",
          "book_id",
          "book_english",
          "book_zomi",
          "chapter",
          "verse",
          "reference",
          "zomi_original",
          "zomi_iso"
        ],
        "title": "BibleVerse",
        "type": "object"
      },
      "DictionaryEntry": {
        "additionalProperties": true,
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "definition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Definition"
          },
          "dialect": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dialect"
          },
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain"
          },
          "english": {
            "title": "English",
            "type": "string"
          },
          "examples": {
            "items": {
              "$ref": "#/components/schemas/TranslateMatch"
            },
            "title": "Examples",
            "type": "array"
          },
          "flagged": {
            "default": false,
            "title": "Flagged",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "letter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Letter"
          },
          "normalized_word": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized Word"
          },
          "part_of_speech": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Of Speech"
          },
          "part_of_speech_full": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Of Speech Full"
          },
          "pronunciation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pronunciation"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "source_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Reference"
          },
          "source_word": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Word"
          },
          "suggested_correction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Correction"
          },
          "usage_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usage Notes"
          },
          "verification_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Status"
          },
          "verified": {
            "default": false,
            "title": "Verified",
            "type": "boolean"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "zomi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zomi"
          }
        },
        "required": [
          "id",
          "english"
        ],
        "title": "DictionaryEntry",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TranslateMatch": {
        "additionalProperties": true,
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain"
          },
          "english": {
            "title": "English",
            "type": "string"
          },
          "exact": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exact"
          },
          "flagged": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flagged"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "part_of_speech": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Of Speech"
          },
          "quality_flags": {
            "items": {
              "type": "string"
            },
            "title": "Quality Flags",
            "type": "array"
          },
          "quality_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Score"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "verification_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Status"
          },
          "verified": {
            "default": false,
            "title": "Verified",
            "type": "boolean"
          },
          "zomi": {
            "title": "Zomi",
            "type": "string"
          }
        },
        "required": [
          "source",
          "english",
          "zomi"
        ],
        "title": "TranslateMatch",
        "type": "object"
      },
      "TranslateRequest": {
        "properties": {
          "direction": {
            "default": "en-zom",
            "pattern": "^(en-zom|zom-en)$",
            "title": "Direction",
            "type": "string"
          },
          "text": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "TranslateRequest",
        "type": "object"
      },
      "TranslationResponse": {
        "additionalProperties": true,
        "properties": {
          "direction": {
            "title": "Direction",
            "type": "string"
          },
          "input": {
            "title": "Input",
            "type": "string"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/TranslateMatch"
            },
            "title": "Matches",
            "type": "array"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "primary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TranslateMatch"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "input",
          "direction",
          "mode",
          "note",
          "matches"
        ],
        "title": "TranslationResponse",
        "type": "object"
      },
      "TranslationSearchResponse": {
        "additionalProperties": true,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/TranslateMatch"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "query",
          "count",
          "results"
        ],
        "title": "TranslationSearchResponse",
        "type": "object"
      },
      "UnifiedSearchResponse": {
        "additionalProperties": true,
        "properties": {
          "bible": {
            "items": {
              "$ref": "#/components/schemas/BibleVerse"
            },
            "title": "Bible",
            "type": "array"
          },
          "dictionary": {
            "items": {
              "$ref": "#/components/schemas/DictionaryEntry"
            },
            "title": "Dictionary",
            "type": "array"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "translations": {
            "items": {
              "$ref": "#/components/schemas/TranslateMatch"
            },
            "title": "Translations",
            "type": "array"
          }
        },
        "required": [
          "query",
          "dictionary",
          "translations",
          "bible"
        ],
        "title": "UnifiedSearchResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Zomi language platform — dictionary, translation retrieval, Bible, and learning.",
    "title": "Siamsil API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/search": {
      "get": {
        "operationId": "unified_search_api_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 8,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unified Search",
        "tags": [
          "search"
        ]
      }
    },
    "/api/translate": {
      "post": {
        "operationId": "translate_api_translate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranslateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Translate",
        "tags": [
          "translate"
        ]
      }
    },
    "/api/translate/search": {
      "get": {
        "operationId": "search_matches_api_translate_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 500,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Matches",
        "tags": [
          "translate"
        ]
      }
    },
...
    "/api/v1/search": {
      "get": {
        "operationId": "unified_search_api_v1_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 8,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unified Search",
        "tags": [
          "search"
        ]
      }
    },
    "/api/v1/translate": {
      "post": {
        "operationId": "translate_api_v1_translate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranslateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Translate",
        "tags": [
          "translate"
        ]
      }
    },
    "/api/v1/translate/search": {
      "get": {
        "operationId": "search_matches_api_v1_translate_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 500,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Matches",
        "tags": [
          "translate"
        ]
      }
    },
...
  }
}